task 3: add a report with dynamic debug overview
This commit contains a report about the investigation of different debug
techniques available in Linux Kernel development. The testing was
performed using the Buildroot environment (QEMU x86_64).

The full report with command listing, output and comments can be found
in the `report.md` file.

The source code of the module is also added (just for reference).

Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>