task 4: add a userspace application to check time
Add a simple userspace application that checks resolution time for all
available system timers and gets the current time.

To allow running the program on different OS (but the same
architecture) it is built with the static linkage.

All helper functions that are used for printing fractional parts are
placed in a separate header file because it is going to be reused for
the kernel module. The main idea behind helper logic is the following:
 - calculate how many zeros have to be prepended to the fractional part
   and prepare a char buffer with them;
 - then remove all trailing zeros;
 - and eventually "concatenate" all parts together.
For simplification proposed to reuse the same char buffer that is
wrapped by the accessor structure. Implementation correctness of memory
operations was checked by the Valgrind util.

Building and running are trivial:
$ make
$ ./main

The program returns the different outputs of time resolution in
different environments:
 - on the host machine:
   CLOCK_REALTIME is ........... 0.000000001
   CLOCK_REALTIME_COARSE is .... 0.004
   CLOCK_MONOTONIC is .......... 0.000000001
   CLOCK_MONOTONIC_COARSE is ... 0.004
   CLOCK_MONOTONIC_RAW is ...... 0.000000001
   CLOCK_BOOTTIME is ........... 0.000000001
   CLOCK_PROCESS_CPUTIME_ID is . 0.000000001
   CLOCK_THREAD_CPUTIME_ID is .. 0.000000001
 - on the emulator:
   CLOCK_REALTIME is ........... 0.004
   CLOCK_REALTIME_COARSE is .... 0.004
   CLOCK_MONOTONIC is .......... 0.004
   CLOCK_MONOTONIC_COARSE is ... 0.004
   CLOCK_MONOTONIC_RAW is ...... 0.004
   CLOCK_BOOTTIME is ........... 0.004
   CLOCK_PROCESS_CPUTIME_ID is . 0.000000001
   CLOCK_THREAD_CPUTIME_ID is .. 0.000000001

Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>