task 4: add a module that returns timestamps
Add a simple Kernel module that interacts with userspace via `sysfs`.
The module has two files:
 - `/sys/kernel/hello/absolute` - for reading the absolute time of a
   previous reading;
 - `/sys/kernel/hello/relative` - for reading elapsed time since the
   previous reading.

To obtain a current time the module uses the function `ktime_get_raw`
that returns the time of the monotonic clock with nanosecond precision.

To convert a `timespec` structure to the string the module uses a set
of helper functions that were implemented earlier for the userspace
application.

Note: the `vmalloc` function is used here to fill callback prototype
requirements (`vmalloc` has the same interface as `malloc`, but
`kmalloc` has an additional parameter). The callback is required by
the helper functions to make them generic and allow using them from
the Kernel and userspace.

According to the task (only six digits should be displayed) the
nanoseconds truncated up to microseconds.

The patch has been successfully tested by:
 $ checkpatch.pl --no-tree -f hello.c

To simplify testing this commit also provides a simple testing script.

Below a possible output:
 # cat /sys/kernel/hello/relative
 22.73882
 # sleep .1; cat /sys/kernel/hello/relative
 0.14556
 # sleep .2; cat /sys/kernel/hello/relative
 0.25604
 # sleep .5; cat /sys/kernel/hello/relative
 0.559808
 # sleep 1; cat /sys/kernel/hello/relative
 1.056131
 # sleep 2; cat /sys/kernel/hello/relative
 2.05554
 # sleep 5; cat /sys/kernel/hello/relative
 5.055936

 # sleep .1; cat /sys/kernel/hello/absolute
 0.000000
 # sleep .2; cat /sys/kernel/hello/absolute
 32.024149
 # sleep .5; cat /sys/kernel/hello/absolute
 32.279498
 # sleep 1; cat /sys/kernel/hello/absolute
 32.836761
 # sleep 2; cat /sys/kernel/hello/absolute
 33.892153
 # sleep 5; cat /sys/kernel/hello/absolute
 35.948182
 # cat /sys/kernel/hello/absolute
 41.003912

Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>